server/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
spacy>=3.7.0
# English NER model (installs as package en-core-web-sm; needed for ENTITY_BACKEND=spacy)
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
anthropic>=0.40.0
httpx>=0.27.0
